Updated rakefile with clean and zip tasks
Added a clean task to delete the build folder and a zip task to archive
the contents of the stage folder to the archive folder.

Also, added support for copying files to the build folder during the build
task and running tests out of the build folder.

Added the total time elapsed if running the default task.

General clean up of the rakefile to parameterize it.